Move all actors into the actors module namespace. Refers to #766
Move Abstract Actors to the Actors Namespace.

Move actors_stack and add_to_collection_actor to the Actors module for namespacing.

Move actors under the actors namespace
