+++ app/actors/curation_concerns/actors/assign_identifier_actor.rb
@@ -1,0 +1,9 @@
+module CurationConcerns
+  module Actors
+    class AssignIdentifierActor < AbstractActor
+      def create(attributes)
+        curation_concern.assign_id && next_actor.create(attributes)
+      end
+    end
+  end
+end
