+++ app/actors/curation_concerns/actors/work_actor_behavior.rb
@@ -1,0 +1,12 @@
+module CurationConcerns
+  module Actors
+    module WorkActorBehavior
+      extend ActiveSupport::Concern
+      extend Deprecation
+
+      included do
+        Deprecation.warn(CurationConcerns::WorkActorBehavior, "CurationConcerns::WorkActorBehavior is deprecated and will be removed in CurationConcerns 1.0")
+      end
+    end
+  end
+end
