--- app/actors/curation_concerns/assign_identifier_actor.rb
@@ -1,7 +1,0 @@
-module CurationConcerns
-  class AssignIdentifierActor < AbstractActor
-    def create(attributes)
-      curation_concern.assign_id && next_actor.create(attributes)
-    end
-  end
-end
