+++ app/services/curation_concerns/actors/actor_factory.rb
@@ -1,0 +1,26 @@
+module CurationConcerns
+  module Actors
+    class ActorFactory
+      def self.build(curation_concern, current_user)
+        Actors::ActorStack.new(curation_concern,
+                               current_user,
+                               stack_actors(curation_concern))
+      end
+
+      def self.stack_actors(curation_concern)
+        [AddToCollectionActor,
+         AssignRepresentativeActor,
+         AttachFilesActor,
+         ApplyOrderActor,
+         InterpretVisibilityActor,
+         model_actor(curation_concern),
+         AssignIdentifierActor]
+      end
+
+      def self.model_actor(curation_concern)
+        actor_identifier = curation_concern.class.to_s.split('::').last
+        "CurationConcerns::#{actor_identifier}Actor".constantize
+      end
+    end
+  end
+end
