+++ lib/generators/curation_concerns/work/templates/actor.rb.erb
@@ -1,6 +1,6 @@
 # Generated via
 #  `rails generate curation_concerns:work <%= class_name %>`
 module CurationConcerns
-  class <%= class_name %>Actor < CurationConcerns::BaseActor
+  class <%= class_name %>Actor < CurationConcerns::Actors::BaseActor
   end
 end
