+++ spec/actors/curation_concerns/interpret_visibility_actor_spec.rb
@@ -1,20 +1,20 @@
 require 'spec_helper'
-describe CurationConcerns::InterpretVisibilityActor do
+describe CurationConcerns::Actors::InterpretVisibilityActor do
   let(:user) { create(:user) }
   let(:curation_concern) { GenericWork.new }
   let(:attributes) { {} }
   subject do
-    CurationConcerns::ActorStack.new(curation_concern,
-                                     user,
-                                     [described_class,
-                                      CurationConcerns::GenericWorkActor])
+    CurationConcerns::Actors::ActorStack.new(curation_concern,
+                                             user,
+                                             [described_class,
+                                              CurationConcerns::GenericWorkActor])
   end
   let(:date) { Date.today + 2 }
 
   describe 'the next actor' do
     let(:root_actor) { double }
     before do
-      allow(CurationConcerns::RootActor).to receive(:new).and_return(root_actor)
+      allow(CurationConcerns::Actors::RootActor).to receive(:new).and_return(root_actor)
     end
 
     context 'when visibility is  set to open' do
