+++ spec/jobs/import_url_job_spec.rb
@@ -28,7 +28,7 @@ describe ImportUrlJob do
     before do
       file_set.id = 'abc123'
       allow(file_set).to receive(:reload)
-      allow(CurationConcerns::FileSetActor).to receive(:new).with(file_set, user).and_return(actor)
+      allow(CurationConcerns::Actors::FileSetActor).to receive(:new).with(file_set, user).and_return(actor)
     end
 
     it 'creates the content' do
@@ -49,7 +49,7 @@ describe ImportUrlJob do
       file_set.save!
       allow(ActiveFedora::Base).to receive(:find).and_call_original
       allow(ActiveFedora::Base).to receive(:find).with(file_set_id).and_return(file_set)
-      allow_any_instance_of(CurationConcerns::FileSetActor).to receive(:create_content)
+      allow_any_instance_of(CurationConcerns::Actors::FileSetActor).to receive(:create_content)
       allow_any_instance_of(Ability).to receive(:can?).and_return(true)
       expect_any_instance_of(Net::HTTP).to receive(:request_get).with(file_hash).and_yield(mock_response)
     end
